pyproject.toml
[project]
name = "strands-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.3",
    "strands-agents-tools>=0.1.2",
]
